Generate RPM package repository metadata with createrepo
This patch sets up to more smoothly support building in an
RHEL 7-based execution environment wherein supplying an RPM
package that provides a createrepo_c executable is challenging.

On CentOS 7 and RHEL 7, the availability of createrepo_c is
uneven, while a createrepo package is more dependable. The
createrepo package supplies only a createrepo executable,
while the createrepo_c package on any platform where it is
supported supplies a createrepo executable.